Rename execution_*_test files to evm_*_test
