+++ test/unittests/CMakeLists.txt
@@ -8,9 +8,10 @@ find_package(GTest CONFIG REQUIRED)
 add_executable(evmone-unittests
     analysis_test.cpp
     bytecode_test.cpp
+    evm_calls_test.cpp
+    evm_fixture.hpp
+    evm_test.cpp
     evmone_test.cpp
-    execution_calls_test.cpp
-    execution_test.cpp
     utils_test.cpp
 )
